fix: inject AgentMetricsRepository in AgentEventHandlerService

+++ backend/src/main/java/dev/backend/sse/controller/dto/MetricEventData.java
@@ -2,7 +2,7 @@ package dev.backend.sse.controller.dto;
 
 public record MetricEventData(
         String timeStamp,
-        Long windowMs,
+        int windowMs,
         Long windowEndMs,
         Long keystrokes
 ) {
